outsourced some types, added framework result to code-generation
