+++ types/package.json
@@ -1,6 +1,6 @@
 {
   "name": "ov-language-server-types",
-  "version": "0.0.1",
+  "version": "0.0.3",
   "description": "type-definitions for the openVALIDATION language-server",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -12,5 +12,8 @@
     "openVALIDATION"
   ],
   "license": "MIT",
-  "readmeFilename": "README.md"
+  "readmeFilename": "README.md",
+  "dependencies": {
+    "typescript": "^3.7.3"
+  }
 }
